+++ test_runner/src/main/kotlin/ftl/args/IArgs.kt
@@ -14,6 +14,7 @@ interface IArgs {
     val recordVideo: Boolean
     val testTimeout: String
     val async: Boolean
+    val clientDetails: Map<String, String>?
     val project: String
     val resultsHistoryName: String?
     val flakyTestAttempts: Int